+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 
 ## Fixed:
 
+- Returning from `game mode` didn't restore user decoration values
 - `remove master` in `master layout` generate LUA runtime error
 - `cava` and `waybar` cava colors weren't syncing with wallpaper
 - Logout on powermenu hanging at black screen